requirements.txt
matplotlib==3.6.2
mysql_connector_repackaged==0.3.1
numpy==1.23.4
pandas==1.5.1
Pillow==9.3.0
streamlit==1.15.2
streamlit_option_menu==0.3.2
